% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.R
\name{glance.lmm}
\alias{glance.lmm}
\title{One-row model summary for an \code{lmm} model}
\usage{
\method{glance}{lmm}(x, ...)
}
\arguments{
\item{x}{An \code{lmm} object.}

\item{...}{Additional arguments.}
}
\value{
A one-row tibble.
}
\description{
One-row model summary for an \code{lmm} model
}
