% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmeans.R
\name{recover_data.lmm}
\alias{recover_data.lmm}
\alias{emm_basis.lmm}
\title{Support for \code{emmeans}}
\usage{
\method{recover_data}{lmm}(object, frame = object$model_frame, ...)

\method{emm_basis}{lmm}(object, trms, xlev, grid, ...)
}
\arguments{
\item{object}{An \code{lmm} object.}

\item{frame}{Model frame.}

\item{...}{Additional arguments passed by \code{emmeans}.}

\item{trms, xlev, grid}{Arguments supplied by \code{emmeans}.}
}
\value{
An \code{emmeans} data-recovery object.

A basis list used by \code{emmeans}.
}
\description{
These methods allow \code{emmeans::emmeans()} to recover the fixed-effects
reference grid and use model-specific Satterthwaite degrees of freedom.
}
