% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sas_line_source}
\alias{sas_line_source}
\title{Line-source irrigation example from the SAS MIXED documentation}
\format{
A data frame with 108 rows and 6 variables:
\describe{
\item{Block}{Block identifier with three levels.}
\item{Cult}{Winter-wheat cultivar with three levels.}
\item{Sbplt}{Ordered subplot position from 1 to 12.}
\item{Irrig}{Irrigation level with six levels.}
\item{Dir}{Side of the sprinkler with levels North and South.}
\item{Y}{Continuous yield response.}
}
}
\source{
SAS Institute Inc. (2017). \emph{SAS/STAT 14.3 User's Guide},
Example 79.6, Line-Source Sprinkler Irrigation.
\url{https://support.sas.com/documentation/onlinedoc/stat/}
}
\usage{
sas_line_source
}
\description{
Winter-wheat yields for three cultivars in three blocks. Each plot contains
twelve subplots arranged on two sides of a line-source sprinkler. SAS uses
these data in Example 79.6 of the SAS/STAT 14.3 documentation to combine
three random-intercept terms with a four-band Toeplitz residual covariance.
}
\keyword{datasets}
