% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sas_split_plot}
\alias{sas_split_plot}
\title{Split-plot example from the SAS MIXED procedure documentation}
\format{
A data frame with 24 rows and 4 variables:
\describe{
\item{Block}{Block identifier with four levels.}
\item{A}{Whole-plot treatment factor with three levels.}
\item{B}{Subplot treatment factor with two levels.}
\item{Y}{Continuous response.}
}
}
\source{
SAS Institute Inc. (2017). \emph{SAS/STAT 14.3 User's Guide},
Example 79.1, Split-Plot Design.
\url{https://support.sas.com/documentation/onlinedoc/stat/}
}
\usage{
sas_split_plot
}
\description{
A balanced split-plot experiment with four blocks, three whole-plot levels,
and two subplot levels. SAS uses this data set in Example 79.1 of the
SAS/STAT 14.3 documentation for the MIXED procedure.
}
\keyword{datasets}
