% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.lmm}
\alias{summary.lmm}
\title{Summarize an \code{lmm} fit}
\usage{
\method{summary}{lmm}(object, level = 0.95, ...)
}
\arguments{
\item{object}{An \code{lmm} object.}

\item{level}{Confidence level for fixed effects.}

\item{...}{Additional arguments.}
}
\value{
An object of class \code{summary.lmm}.
}
\description{
Summarize an \code{lmm} fit
}
