% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.R
\name{tidy.lmm}
\alias{tidy.lmm}
\title{Tidy an \code{lmm} model}
\usage{
\method{tidy}{lmm}(
  x,
  effects = c("fixed", "ran_pars", "all"),
  conf.int = FALSE,
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{x}{An \code{lmm} object.}

\item{effects}{Effects to return: fixed effects, covariance parameters, or
both.}

\item{conf.int}{Whether to include confidence limits for fixed effects.}

\item{conf.level}{Confidence level.}

\item{...}{Additional arguments.}
}
\value{
A tibble.
}
\description{
Tidy an \code{lmm} model
}
