% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{VarCorr}
\alias{VarCorr}
\title{Extract covariance components}
\usage{
VarCorr(object, ...)
}
\arguments{
\item{object}{A fitted model.}

\item{...}{Additional arguments.}
}
\description{
Extract covariance components
}
