% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.R
\name{augment.lmm}
\alias{augment.lmm}
\title{Augment data with fitted values and residuals}
\usage{
\method{augment}{lmm}(x, data = NULL, newdata = NULL, ...)
}
\arguments{
\item{x}{An \code{lmm} object.}

\item{data}{Optional data used for augmentation. By default, excluded rows
are restored when the model used \code{\link[stats:na.fail]{stats::na.exclude()}}.}

\item{newdata}{Optional new data. If supplied, it takes precedence over
\code{data}.}

\item{...}{Additional arguments passed to \code{\link[=predict]{predict()}}.}
}
\value{
A tibble with \code{.fitted}, \code{.resid}, and \code{.std.resid} columns.
}
\description{
Augment data with fitted values and residuals
}
